+++ src/iwstring.h
@@ -1,6 +1,7 @@
 #ifndef IWSTRING_H
 #define IWSTRING_H
 
+#include <cstddef>
 #include <cstring>
 #include <vector>
 #include <ostream>
@@ -23,13 +24,13 @@ public:
     String& operator+=(const String& other);
     bool operator<(const String& other);
     friend std::ostream& operator<<(std::ostream& os, const String& other);
-    char& operator[](const int& index);
+    char& operator[](const size_t& index);
     bool operator==(const String& other);
     bool operator!=(const String& other);
 
     String& append(const String& other);
 
-    char& characterAt(int index) const;
+    char& characterAt(size_t index) const;
 
     String& clear();
 
@@ -43,15 +44,15 @@ public:
     void setData(const char* d, bool clearFirst);
     void setData(const String& str, bool clearFirst);
 
-    const int find(const String& findStr) const;
-    const int find(const int& startIndex, const String& findStr) const;
+    const size_t find(const String& findStr) const;
+    const size_t find(const size_t& startIndex, const String& findStr) const;
 
-    const int length() const;
-    const int length(const bool includeNullTerminator) const;
-    static const int length(const char* str);
-    static const int length(const char* str, const bool includeNullTerminator);
-    static const int length(const String& str);
-    static const int length(const String& str, const bool includeNullTerminator);
+    const size_t length() const;
+    const size_t length(const bool includeNullTerminator) const;
+    static const size_t length(const char* str);
+    static const size_t length(const char* str, const bool includeNullTerminator);
+    static const size_t length(const String& str);
+    static const size_t length(const String& str, const bool includeNullTerminator);
 
     String& prepend(const String& other);
